Extracted visualization of CiRA functionality into specific component
